Upped Minor -> V03.03.09
Created the Json Schema for migration from JAXB to Jackson.
The Json Schema was created mimicking the SWAMP schema.
Since Jackson can handle XML relatively easily, it was chosen as the\
new library.

Unit tests are passed, but they are not thorough tests.
The three tests within the Makefile do pass and get validated against\
the SWAMP schema.

Documentation needs to be heavily done, as references for the creation\
need to be properly attributed.

Could also combine the serializers into a single path?

Signed-off-by: franceme <frantzme@vt.edu>