Upped Minor -> V03.04.00
Prepped for release.

Added the Utils projectVersion into the autoreplace, as loading the\
properties file dynamically was not reliable.
Due to this the @since version needed to change to help differentiate.

Also didn't include the tests within the building of the makefile since\
the gradle version may need to use Java8, however this is still\
dependent on Java7.

Signed-off-by: franceme <frantzme@vt.edu>