Upped Minor Version -> V03.04.04.
Reworded and refreshed the README.
Removed the Makefile.rst.
Enhanced the Makefile, provided info messages, silenced certain output\
and added steps to verify the environment variables.
Updated the Scarf schema.
Made the output more useful, (the help and the version).\
It now doesn't act like an error is being thrown.

Signed-off-by: franceme <9311496+franceme@users.noreply.github.com>